// app/utils/types.ts
export type LayoutOption = {
    key: string;
    label: string;
    width: number;
    height: number;
};

export type MarginOption = {
    key: string;
    label: string;
    top: number;
    bottom: number;
    left: number;
    right: number;
};

export type FormField = {
    label: string;
    key: string;
    value: string;
};

export type Pagination = {
    id: number;
    content: string;
};